Move autocomplete from timetable script to `SearchBar` component
